requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
ccxt==3.0.56
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
PyQt5==5.15.9
PyYAML==6.0
python-dotenv==1.0.0
